Isolate mod-2 zero-half cut generation
Move the separator into a dedicated compilation unit and retain candidate parity data without duplicate storage.
